+++ R/calibrate.R
@@ -605,27 +605,25 @@ draw.pbmodelled <- function(set=get('info'), BCAD=set$BCAD, rotate.axes=FALSE, r
     this <- ifelse(rotate.axes, 3, 4)
     pretty.pb <- pretty(c(pbmin, pbmax)) # not OK?
     pretty.pb <- pretty(pb.lim)
-    if(BCAD) {
-      onad <- pb2ad(pretty.pb)
-      axis(this, rev(onad), rev(pretty.pb), col=pbmeasured.col, col.axis=pbmeasured.col, col.lab=pbmeasured.col)
-    } else {
-        onbp <- pb2bp(pretty.pb)
+    onbp <- pb2bp(pretty.pb)
+    if(BCAD)
+      axis(this, rev(onbp), rev(pretty.pb), col=pbmeasured.col, col.axis=pbmeasured.col, col.lab=pbmeasured.col) else
         axis(this, onbp, pretty.pb, col=pbmeasured.col, col.axis=pbmeasured.col, col.lab=pbmeasured.col)
-      }
-    mtext(pb.lab, this, 1.7, col=pbmeasured.col, cex=.8)
+
+    mtext(pb.lab, this, 1.8, col=pbmeasured.col, cex=.8)
 
     for(i in 1:length(depths)) {
-      if(BCAD)
-        ages <- pb2ad(rev(Ai$x[[i]])) else
+      if(BCAD) {
+        ages <- pb2bp(rev(Ai$x[[i]]))
+        z <- t(rev(Ai$y[[i]]))/hght
+      } else {
           ages <- pb2bp(Ai$x[[i]])
+          z <- t(Ai$y[[i]])/hght
+        }
 
-    if(BCAD)
-      z <- t(rev(Ai$y[[i]]))/hght else
-        z <- t(Ai$y[[i]])/hght
-
-    if(rotate.axes)
-      image(ages, c(depths[i]-thickness[i], depths[i]), z, col=pbmodelled.col(seq(0, 1-max(z), length=50)), add=TRUE) else
-        image(c(depths[i]-thickness[i], depths[i]), ages, z, col=pbmodelled.col(seq(0, 1-max(z), length=50)), add=TRUE)
+      if(rotate.axes)
+        image(ages, c(depths[i]-thickness[i], depths[i]), z, col=pbmodelled.col(seq(0, 1-max(z),  length=50)), add=TRUE) else
+          image(c(depths[i]-thickness[i], depths[i]), ages, z, col=pbmodelled.col(seq(0, 1-max(z), length=50)), add=TRUE)
     }
   }
 
@@ -657,7 +655,7 @@ draw.pbmodelled <- function(set=get('info'), BCAD=set$BCAD, rotate.axes=FALSE, r
 #' @author Maarten Blaauw
 #' @return a list of modelled values of A
 #' @export
-A.modelled <- function(d.top, d.bottom, dens, set=get('info'), phi=set$phi, sup=set$ps, multiply=1) {
+A.modelled <- function(d.top, d.bottom, dens, set=get('info'), phi=set$phi, sup=set$ps) {
   if(d.top >= d.bottom)
     stop("\n d.top should be above d.bottom", call.=FALSE)
   dd <- 1
@@ -670,6 +668,7 @@ A.modelled <- function(d.top, d.bottom, dens, set=get('info'), phi=set$phi, sup=
   t.top <- Bacon.Age.d(d.top, BCAD=F) - set$theta0
   t.bottom <- Bacon.Age.d(d.bottom, BCAD=F) - set$theta0
   #  multiply <- ifelse(set$Bqkg, 10, 500)
+  multiply <- 1 # since Bqkg or dpmg is already set earlier (for set$dets and set$detsPlum)
   return(sup + ((phi / (.03114*multiply*dens) ) * (exp( -.03114*t.top) - exp(-.03114*t.bottom)) ) )
 }